docs(changeset): Erstatter intern Calendar med den eksternt tilgjengelige
