+++ .changeset/mighty-plants-film.md
@@ -1,0 +1,5 @@
+---
+"@fremtind/jokul": patch
+---
+
+Flytter skillelinjen mellom to input-feltet til en delt mixin, og legger til fix for overlay i Safari
