+++ .changeset/young-places-cover.md
@@ -1,0 +1,5 @@
+---
+"@fremtind/jokul": minor
+---
+
+Erstatter intern Calendar med den eksternt tilgjengelige
