--- packages/jokul/src/components/calendar/styles/_index.scss
@@ -1,1 +1,0 @@
-@forward "calendar";
