--- packages/jokul/src/components/datepicker/calendar/index.ts
@@ -1,3 +1,0 @@
-export { Calendar } from "./Calendar.js";
-
-export type { CalendarProps } from "./types.js";
