+++ packages/jokul/src/components/datepicker/calendar/styles/_index.scss
@@ -1,1 +1,2 @@
 @forward "calendar";
+@forward "calendar-day";
