+++ packages/jokul/src/components/datepicker/date-input/styles/_index.scss
@@ -1,0 +1,1 @@
+@forward "date-input";
