+++ packages/jokul/src/components/datepicker/development/DatepickerExample.tsx
@@ -3,10 +3,17 @@ import type {
     ExampleComponentProps,
     ExampleKnobsProps,
 } from "utils/dev-example/index.js";
+import { Help } from "../../help/index.js";
 import type { LabelVariant } from "../../input-group/types.js";
-import { PopupTip } from "../../tooltip/PopupTip.js";
-import { DatePicker } from "../DatePicker.js";
-import { formatInput, isBlurTargetOutside } from "../utils.js";
+import { DatePicker } from "../date-picker/DatePicker.js";
+
+const today = new Date();
+const monthsPast = new Date(
+    new Date(today).setMonth(new Date(today).getMonth() - 14),
+).toISOString();
+const monthsComing = new Date(
+    new Date(today).setMonth(new Date(today).getMonth() + 14),
+).toISOString();
 
 export const datepickerExampleKnobs: ExampleKnobsProps = {
     boolProps: ["Med hjelpetekst", "Med feil", "Med tooltip"],
@@ -17,11 +24,6 @@ export const datepickerExampleKnobs: ExampleKnobsProps = {
             defaultValue: 0,
         },
     ],
-};
-
-const monthsIsh = (num: number) => {
-    const raw = 1000 * 60 * 60 * 24 * (num * 30 - 5);
-    return raw < 0 ? 12 + raw : raw;
 };
 
 export const DatepickerExample: FC<ExampleComponentProps> = ({
@@ -37,10 +39,13 @@ export const DatepickerExample: FC<ExampleComponentProps> = ({
     const variant = choiceValues && (choiceValues["Variant"] as LabelVariant);
 
     const tooltip = boolValues?.["Med tooltip"] ? (
-        <PopupTip content="Du vil være forsikret fra denne datoen. Du kan ikke velge en dato som har vært." />
+        <Help buttonText="Hjelp">
+            Du vil være forsikret fra denne datoen. Du kan ikke velge en dato
+            som har vært.{" "}
+        </Help>
     ) : undefined;
 
-    const [value, setValue] = useState<string>("");
+    const [value, setValue] = useState<string>("2002-20-20");
 
     return (
         <DatePicker
@@ -50,50 +55,24 @@ export const DatepickerExample: FC<ExampleComponentProps> = ({
             errorLabel={errorLabel}
             name="datepicker"
             helpLabel={helpLabel}
-            disableBeforeDate={formatInput(
-                new Date(Date.now() - monthsIsh(14)),
-            )}
-            disableAfterDate={formatInput(new Date(Date.now() + monthsIsh(14)))}
+            min={monthsPast}
+            max={monthsComing}
             value={value}
-            onFocus={(e, date, meta) => {
+            onFocus={(e) => {
                 console.log("onFocus", {
                     event: e,
-                    date,
-                    meta,
                 });
             }}
-            onBlur={(e, date, meta) => {
-                // Ignorer blurs som går til kalenderknapper
-                if (isBlurTargetOutside(e)) {
-                    console.log("onBlur", {
-                        event: e,
-                        date,
-                        meta,
-                    });
-                }
+            onBlur={(e) => {
+                console.log("onBlur", {
+                    event: e,
+                });
             }}
-            onChange={(e, date, meta) => {
+            onChange={(e) => {
                 setValue(e.target.value);
                 console.log("onChange", {
                     event: e,
-                    date,
-                    meta,
                 });
-            }}
-            action={{
-                onBlur: (e) => {
-                    // Ignorer blurs som går tilbake til inputfeltet
-                    if (isBlurTargetOutside(e)) {
-                        console.log("action.onBlur", {
-                            event: e,
-                        });
-                    }
-                },
-                onClick: (e) => {
-                    console.log("action.onClick", {
-                        event: e,
-                    });
-                },
             }}
         />
     );
