--- packages/jokul/src/components/datepicker/styles/_calendar-navigation.scss
@@ -1,14 +1,0 @@
-@use "../../../styles/jkl";
-
-@layer jokul.components {
-    .jkl-calendar-navigation {
-        border: 0;
-        padding: 0;
-        margin: 0;
-        display: flex;
-        box-sizing: border-box;
-        width: 100%;
-        justify-content: space-between;
-        align-items: center;
-    }
-}
