--- packages/jokul/src/components/datepicker/styles/_calendar-table.scss
@@ -1,29 +1,0 @@
-@use "../../../styles/jkl";
-
-@layer jokul.components {
-    .jkl-calendar-table {
-        // Calendar table heading (weekday names)
-        // ======================================
-
-        th {
-            @include jkl.text-style("text-small");
-            padding-bottom: var(--jkl-unit-10);
-        }
-
-        // Calendar table styles
-        // =====================
-
-        td {
-            text-align: center;
-        }
-
-        td + td,
-        th + th {
-            padding-left: calc(var(--jkl-unit-10) / 2);
-        }
-
-        tr + tr > td {
-            padding-top: calc(var(--jkl-unit-10) / 2); // space between rows
-        }
-    }
-}
