--- packages/jokul/src/components/datepicker/styles/_index.scss
@@ -1,6 +1,0 @@
-@forward "datepicker";
-@use "../../button/styles" as button;
-@use "../../icon/styles" as icon;
-@use "../../input-group/styles" as input-group;
-@use "../../text-input/styles" as text-input;
-@use "../../popover/styles" as popover;
