+++ packages/jokul/src/styles/components.scss
@@ -9,7 +9,8 @@
 @use "../components/combobox/styles" as combobox;
 @use "../components/cookie-consent/styles" as cookie-consent;
 @use "../components/countdown/styles" as countdown;
-@use "../components/datepicker/styles" as datepicker;
+@use "../components/datepicker/date-input/styles" as date-input;
+@use "../components/datepicker/date-picker/styles" as datepicker;
 @use "../components/description-list/styles" as description-list;
 @use "../components/expander/styles" as expander;
 @use "../components/feedback/styles" as feedback;
